--- src/components/Modal/CreateTask.js
@@ -1,18 +1,0 @@
-import api from './../../shared/service/axios/axiosClient';
-
-export const CreateTask = ({ description, title, results }) => {
-  api.post(
-    '/tasks',
-    {
-      description,
-      title,
-      results,
-    },
-    {
-      headers: {
-        Authorization:
-          'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2ZTIyZjU1ZS0zZjFiLTQ0YzctODdkNi1hYjE0OGQ5ODM0MmYiLCJ1c2VybmFtZSI6InRlYW0wIiwiaWF0IjoxNjk3NjIxNTg3LCJleHAiOjE2OTc3MDc5ODd9.fofkVpMGEm0awjsOLTam7gDOx0BJS1nfi4iBvkD-3sg',
-      },
-    },
-  );
-};
