+++ src/entities/TaskCard/index.ts
@@ -1,0 +1,1 @@
+export { TaskCard } from './TaskCard';
