+++ src/entities/UserApi/user.interface.ts
