+++ src/features/AnswerForm/index.ts
@@ -1,0 +1,1 @@
+export { AnswerForm } from './AnswerForm';
