+++ src/pages/TaskPage/index.ts
@@ -1,0 +1,1 @@
+export { TaskPage } from './TaskPage';
