+++ src/pages/TasksPage/TasksPage.jsx
@@ -1,19 +1,17 @@
-import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import ModalCustom from "../../components/Modal/Modal";
-import Button from "@mui/material/Button";
-import TaskCard from "../../components/TaskCard";
-import api from "../../shared/service/axios/axiosClient";
-import Typography from "@mui/material/Typography";
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { CreateTaskModal } from '../../widgets/CreateTaskModal';
+import Button from '@mui/material/Button';
+import { TaskCard } from '../../entities/TaskCard';
+import api from '../../shared/service/axios/axiosClient';
+import Typography from '@mui/material/Typography';
 
-import Loading from "../../components/Loading";
-import { noData } from "./tasksPage.model";
-import Header from "../../components/Header/Header";
-import { useAuth } from "../../shared/hooks/useAuth";
-import userService from "../../shared/services/user/user.service";
-import taskService from "../../shared/services/task/task.service";
+import { Loading } from '../../shared/components/Loading';
+import { noData } from './tasksPage.model';
+import { Header } from '../../widgets/Header';
+import { useAuth } from '../../shared/hooks/useAuth';
 
-const TasksPage = () => {
+export const TasksPage = () => {
   const [open, setOpen] = useState(false);
   const [tasks, setTasks] = useState([]);
   const navigate = useNavigate();
@@ -23,7 +21,7 @@ const TasksPage = () => {
   const handleOpen = () => setOpen(true);
 
   const handleClick = (id) => {
-    navigate("/tasks/" + id);
+    navigate('/tasks/' + id);
   };
 
   const GetTasks = () => {
@@ -39,50 +37,20 @@ const TasksPage = () => {
     }
   }, [isLoading]);
 
-  // const getUser = async () => {
-  //   const data = await userService.getUser();
-  //   console.log(data);
-  // };
-
-  // const editUser = async (data) => {
-  //   await userService.editUser(data);
-  // };
-
-  // const getConnectUser = async (id) => {
-  //   const data = await userService.getConnectUser(id);
-  //   console.log(data);
-  // };
-
-  // const createTask = async (data) => {
-  //   await taskService.createTask(data);
-  //   console.log(data);
-  // };
-
-  // const getTaskById = async (id) => {
-  //   const data = await taskService.getTaskById(id);
-  //   console.log(data);
-  // };
-
   return (
     <div>
-      {/* <button
-        onClick={() => getTaskById("e7cd0d2a-9b37-4809-9c7f-5db2d37e7886")}
-      >
-        qwe
-      </button> */}
-      {/* <button onClick={() => getUser()}>qwe</button> */}
       {isLoading && <Loading />}
       <Header onClick={handleOpen} />
       <div
         style={{
-          display: "flex",
-          flexWrap: "wrap",
-          gap: "20px",
-          padding: "16px 0",
+          display: 'flex',
+          flexWrap: 'wrap',
+          gap: '20px',
+          padding: '16px 0'
         }}
       >
         {(!tasks.data || tasks.data.length === 0) && (
-          <Typography variant="h5" component="div">
+          <Typography variant='h5' component='div'>
             {noData}
           </Typography>
         )}
@@ -98,9 +66,7 @@ const TasksPage = () => {
             );
           })}
       </div>
-      <ModalCustom taskFoo={GetTasks} open={open} close={setOpen} />
+      <CreateTaskModal taskFoo={GetTasks} open={open} close={setOpen} />
     </div>
   );
 };
-
-export default TasksPage;
