+++ src/pages/TasksPage/index.ts
@@ -1,0 +1,1 @@
+export { TasksPage } from './TasksPage';
