+++ src/shared/components/InputCustom/index.ts
@@ -1,0 +1,1 @@
+export { InputCustom } from './InputCustom';
