+++ src/shared/components/Loading/index.ts
@@ -1,0 +1,1 @@
+export { Loading } from './Loading';
