+++ src/widgets/CreateTaskModal/index.ts
@@ -1,0 +1,1 @@
+export { CreateTaskModal } from './CreateTaskModal';
