+++ src/widgets/Header/assets/user.png
