+++ src/widgets/TimerCustom/index.ts
@@ -1,0 +1,1 @@
+export { TimerCustom } from './TimerCustom';
